pyproject.toml
[project]
name = "mail-agent"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ashutosh Kumar", email = "ashutosh@combinehealth.ai" }
]
dependencies = [
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.164.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "imaplib2>=2.57.2",
    "pyyaml>=6.0",
    "schedule>=1.2.2",
    "beautifulsoup4>=4.12.2",
    "nltk>=3.8.1",
    "scikit-learn>=1.3.0",
    "pandas>=2.0.3",
    "email-validator>=2.0.0",
    "notifiers>=1.3.3",
    "twilio>=8.5.0",
    "openai>=1.66.3",
    "anthropic>=0.49.0",
    "setuptools>=76.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/mail_agent"]
